Makefile: fix encoding issue for dumpdb
- fix: fix so that "make dumpdb" works for "manage.py dump"
Issue: "make dumpdb" was giving the following error when it ran "manage.py dump":
    UnicodeEncodeError: 'ascii' codec can't encode character u'\xd7' in position 800: ordinal not in range(128)
Setting the PYTHONIOENCODING=utf-8 environment variable fixes this.
Note that it worked if there wasn't a redirect to a file, because then it used the console's encoding.
But when it was redirected to the file, then there was no encoding, so it used 'ascii'.