UPGRADE to Django 3: FIX question page error with unexpected keyword argument 'field_name' (WORKING)
With the upgrade to Django 3, the filter().latest() method has the removed.
Fixed this by removing the keyword argument and just using a positional argument.

Error that this fixes:
    Get the following error when hitting the http://127.0.0.1:8000/question/ page:
      File "/Users/rbednark/quizme-read-only-prod/questions/views.py", line 377, in question
        ).latest(field_name='datetime_added')
      TypeError: latest() got an unexpected keyword argument 'field_name'

NOTE: this commit works, and is the first one working since the Django 3 upgrade.
However, note that the formatting on the question page has an issue
where the attempt input box is pushed to the right and overlaps the tags.