UPGRADE django-pagedown to 2.1.2: questions.forms.FormAttemptNew: specify css file in class
ISSUE: the text box for the answer on the question/attempt page is
pushed to the right, and overlaps the tags.

FIX:
With the upgrade of django-pagedown from 1.0.4 to 2.1.2, the custom
css that is needed for the question page needs to be specified in
an inherited class, whereas before I think it just needed to be
present in a pagedown/custom.css file in the templates path.

NOTE: this completes the upgrade to Django 3, along with all other
packages upgraded to their latest versions (except splinter).
Everything is now working.