FIX questions/views.py handle DoesNotExist exception
ISSUE: if there is no question to show for the tags selected, and then
the user selects different tags and clicks "submit", a DoesNotExist
exception is raised.

Fix: if there is no question to be shown for the answer page, catch
the exception, and redirect back to the question page, because the user may
have selected different tag(s).