COMMENT README TODO: remove completed item
Remove item for feature to put question and response in a single page,
because this has been implemented.

Remove item about seeing if I can move db data into a fresh db.
I tried this, and it is not easy -- need to manually modify data for
tables used by Django.  See TESTING.md