FIX flashcard.html change <details open=true> to <details open>
ISSUE: the <details> "open" attribute was raising an HTML warning, because it should not take a value
The "open" attribute is boolean -- true if it exists, false if it doesn't exist.  It does not take a value.

FIX: remove the "=true" value for the "open" attribute

No change in behavior.