FIX: _get_oldest_viewed_tag(): the oldest (Least()) date was being used instead of the newest (Greatest())
Also:

• removed Coalesce()
It's not needed, because the When() condition has values for both variables.
However, consider using it and removing the Case..When, as I think that would make it simpler,
and might improve performance.

• Fixed tests: expected values for QUERY_*_BY_TAG tests
Some of the expected values were wrong.  I was confused in determining what
the oldest tag is.